+++ _common/gameModel/libs/ClientScript/game/common/actions/actionTemplate/mapTemplates.ts
